Add sidebar to vue sheet, layout improvements
