+++ src/module/sheets/_vue/_base-vue-actor-sheet.mjs
@@ -13,7 +13,7 @@ export class GrimwildBaseVueActorSheet extends foundry.applications.sheets.Actor
 		viewPermission: DOCUMENT_OWNERSHIP_LEVELS.LIMITED,
 		editPermission: DOCUMENT_OWNERSHIP_LEVELS.OWNER,
 		position: {
-			width: 600,
+			width: 800,
 			height: 600
 		},
 		actions: {
