+++ src/vue/components/index.mjs
@@ -1,4 +1,6 @@
 // Character.
+export { default as CharSidebar } from '@/components/actor/character/CharSidebar.vue';
+export { default as CharHeader } from '@/components/actor/character/CharHeader.vue';
 export { default as CharDetails } from '@/components/actor/character/CharDetails.vue';
 export { default as CharTalents } from '@/components/actor/character/CharTalents.vue';
 export { default as CharEffects } from '@/components/actor/character/CharEffects.vue';
